src/requirements.txt
pandas
SQLAlchemy
PyMySQL==1.0.2
kafka-python==2.0.2
Faker